requirements.txt
line-bot-sdk
beautifulsoup4
requests
openai
reportlab
pandas
